Leader board modal fixes, integrate wallet api with game hub for game lobby enter & leave (#43)
* add github pipeline

* docker image pipeline fixes

* Game lounge fixes

* Fixes, Game Lobby Join, GameLobby Przies modal

* UI Fixes

* profile updates, user achievements, user game play history

* fixes

* Rebuild full game lobby

* fixes

* wip

* WIP 👷‍♂️

* WIP 👷‍♂️

* WIP 👷‍♂️

* WIP 👷‍♂️

* WIP 👷‍♂️

* fixes

* fixes

* fixes

* WIP 👷‍♂️

* fixes

* fixes

* more demo users

* add more users

* update README.md

* empty line

* 88 - Cooldown after leaving game lobby (#4)

* 243 - user assets overview & assets page (#5)

* implementing wallete manegmeng task

made the quary and  edited the dashboard vue and mae assetaccount vue for view made the route and the assetaccount invocable controoler

* asset account pipkine

some clean up and made the pipline and  the filters

* finished user wallet profile

finished user wallet table  on profile page

* finished vue Asset account page

asset account page with filtering and pagination

* fixed a bug

switched the achievement with asset account

* fixes

fixing asset account page

* fixes

fixing the data sent to asset account page

* clean up

deleted unneeded data

* finalize assets page + fix merge conflict


* Refactoring and fixes

* update .env.example

* Notifications API and UI (#3)

* fixing bug

fixing notifications service bug

* after review

committing changes after review

* convert notification id to uui

* finished Notification service backend

finished the notification service backend and tested it manually

* fixes

* Notification fixes

* making the notification UI

made the controllers the routes and the vue component with its function

* added pagination for notification service

add new navegation component and built the pagination

* fix

fix the ui when there is more than 5 elements

* add-result-notification-of-gamelobby (processing || processed)

* Fixes modal

* fix currentUser


* fix inertia middleware

* Add sign out button to default template

* finished the pagination for game lobby (#9)

* Add wodoland, wodo shooter

* fixes

* Add wodoland, wodo shooter

* agario

* change game=lobby placeholder image

* finished the transaction tabel without sorting (#6)

* finished the transaction tabel without sorting

did the controller  and transaction resource ,piptline , transaction table

* finished the pipeline and some clean up

made all the filters and cleaned up the code

* changed the transaction dummy data to api call

* fixed the api pagenation

made a pagination based on the api


* Finalizing transactions

* Yamanjamal/40 withdrow screen (#7)

* finished the transaction tabel without sorting

did the controller  and transaction resource ,piptline , transaction table

* started with withdrawal transaction

did the table and the controller , route, and the table

* finished withdraw screen ui

finished the form of withdraw as a ui

* Update Withdraw.vue

* change the transaction dummy data to api

fixed the withdrawal table and change the dummy data to api

* fixed the pagination of withdrawal


* Yamanjamal/44 deposit screen (#10)

* finished the transaction tabel without sorting

did the controller  and transaction resource ,piptline , transaction table

* started with withdrawal transaction

did the table and the controller , route, and the table

* finished withdraw screen ui

finished the form of withdraw as a ui

* Update Withdraw.vue

* finished Deposit ui

did the ui of deposit screen with route and controller

* made the pipelind and some clean up

made the boiler plate for pipeline

* made the pipelines filters for deposit and withdrawal

some clean up and made the pipeline for deposit and withdrawal

* changed the deposit dummy data to api

dummy to api and made the table of deposit

* fixed the pagination of deposit


* fix

* Fixes

* fixes

* WIP 👷‍♂️

* fix image

* WIP 👷‍♂️

* add withdraw and deposite controller

* Update main logo

* Withdraw, Deposit, transactions Filters (#11)

* add profile dropdown

* Yamanjamal/206 gaming experience initiate game lobby (#12)

* finished the admin games table

middleware, routes, pipeline, controllers, the view admin games page, game lobby table, enums, gamehub lang

* finished the create and add functions and fixed the pipelines

* finished the admin Gamelobby CRUD feature

* clean up


* Sorting fixes user games played history pipeline

* fix all enums

* add admin button

* add admin button - hide for non-admins

* Wallet enums fixes

* add redeploy scropt and some fixes for admin query pipeline

* Fix renaming

* Fixes

* Fixes

* fix redeploy script

* crud gameLobbyTemplate feature fdone (#15)

* crud gameLobbyTemplate feature fdone

* lobby and template table fixes for the search

* Yamanjaml/258-transaction-logs (#14)

* finished most of the ui

* finished the transaction dialog ui

* some fixes

* fixing the api call

* small clean up

* the transaction log feature is done

* some clean up (#16)


* finished profile branch (#13)


* Danyel/fix seeding (#17)

* temporal setup

* Fix seedings and some UI

* temporal prize distribution, UI Fix and seed update

* Fix withdraw and Transaction UI

* Yamanjamal/251 withdrawal with verification code sent via email (#18)

* finished the ui and the controllers and the migrations and the request

* finished the feature

* change driver from database to mail

* done fixing the admin controllers structure (#19)

* Yamanjamal/UI fixes and clean up (#20)

* some ui fixes that need to be tested

* finished the test for ui fixes


* some route changes (#21)


* logout fix

* fix gamehub.wodo.io

* fix gamehub.wodo.io

* support WSS for sockets

* change file name

* Deploy Temporal Workers Docker image

* Add new notification types + authorize sending post notificaitons

* add notifcaiton UI

* fix notifications

* temporal machine to machine resource fixes

* fixing the leaderboard list pulling (#23)

* editing the leaderboard winners model

* new gamelobby current user api

* Fix leaderboard modal issues, design and appearing


* Yamanjamal/replace the account increment and decrement to api call (#22)

* got the code ready just need to configure the api

* refid fixed

* done but need testing

* fix seeded data

* fixing the api end point

* replacing the db user asset account with the api end point

* finished adding get user asset accounts action

+++ app/Actions/Games/GameLobbies/RemoveUserFromGameLobbyAction.php
@@ -2,21 +2,27 @@
 
 namespace App\Actions\Games\GameLobbies;
 
-use App\Enums\Reactions\RemoveUserFromGameLobbyReaction;
-use App\Events\GameLobby\PrizeUpdatedEvent;
-use App\Events\GameLobby\UserLeftGameLobbyEvent;
-use App\Models\ChatRoomUser;
-use App\Models\GameLobby;
-use App\Models\GameLobbyUser;
-use App\Models\WodoAssetAccount;
+use DB;
 use Auth;
 use Cache;
-use DB;
 use Event;
+use App\Models\GameLobby;
+use App\Models\ChatRoomUser;
 use Illuminate\Http\Request;
+use App\Models\GameLobbyUser;
+use App\Models\WodoAssetAccount;
+use Illuminate\Support\Facades\Http;
+use App\Events\GameLobby\PrizeUpdatedEvent;
+use App\Events\GameLobby\UserLeftGameLobbyEvent;
+use App\Actions\Wallet\GetUserAssetAccountAction;
+use App\Enums\Reactions\RemoveUserFromGameLobbyReaction;
 
 class RemoveUserFromGameLobbyAction
 {
+    public function __construct(public GetUserAssetAccountAction $getUserAssetAccountAction)
+    {
+    }
+
     public function execute(Request $request, GameLobby $gameLobby): GameLobby|RemoveUserFromGameLobbyReaction
     {
         $user = $request->user();
@@ -34,12 +40,16 @@ class RemoveUserFromGameLobbyAction
                 ) {
                     return RemoveUserFromGameLobbyReaction::UserNotInGameLobby;
                 }
+                //geting the user account from the api end point
+                $userAssetAccount = $this->getUserAssetAccountAction->execute($gameLobby->asset);
 
-                $userAssetAccount = $user
-                    ->assetAccounts()
-                    ->lockForUpdate()
-                    ->where('asset_id', $gameLobby->asset_id)
-                    ->first();
+                // dd($userAssetAccount->id)
+
+                // $userAssetAccount = $user
+                //     ->assetAccounts()
+                //     ->lockForUpdate()
+                //     ->where('asset_id', $gameLobby->asset_id)
+                //     ->first();
 
                 $gameLobbyUser = $gameLobby
                     ->users()
@@ -50,19 +60,42 @@ class RemoveUserFromGameLobbyAction
                 $gameLobbyUserEntranceFee = $gameLobbyUser->pivot->entrance_fee;
 
                 // return the amount he paid
-                $userAssetAccount->increment('balance', $gameLobbyUserEntranceFee);
+                //here we should call the api
+                $url = config('wodo.wallet-transactions-api') . 'home-withdraw';
+                $data = [
+                    'toAccountId' => $userAssetAccount->id,
+                    'asset' => $gameLobby->asset->symbol,
+                    'amount' => $gameLobby->base_entrance_fee,
+                    'refId' => $gameLobby->id,
+                ];
+                
+                $response = Http::post(url: $url, data: $data);
+                
+                if ($response->failed()) {
+                    return $response->toException();
+                }
+                // dd($userAssetAccount->balance);
 
-                $wodoAccount = WodoAssetAccount::sharedLock()
-                    ->where('asset_id', $gameLobby->asset_id)
-                    ->first();
 
-                $wodoAccount->decrement('balance', $gameLobbyUserEntranceFee);
+                // return $response->body();
+
+                // $userAssetAccount->increment('balance', $gameLobbyUserEntranceFee);
+
+                // $wodoAccount = WodoAssetAccount::sharedLock()
+                //     ->where('asset_id', $gameLobby->asset_id)
+                //     ->first();
+
+                // $wodoAccount->decrement('balance', $gameLobbyUserEntranceFee);
 
                 $gameLobby->increment('available_spots');
 
                 $gameLobby->users()->detach([$user->id]);
 
                 ChatRoomUser::where([['chat_room_id', '=', $gameLobby->id], ['user_id', '=', $user->id]])->delete();
+
+                //forgeting the user asset account after the balance changed
+                Cache::forget('user.' . Auth::id() . '.account' . $gameLobby->asset->symbol);
+                Cache::forget('user.' . Auth::id() . '.accounts');
 
                 Cache::forget('user.' . Auth::id() . '.current-lobby-session');
